For SVY-19308: Add general docs to API nodes that are implemented in java
As not all our code comments contain descriptions that are consistent,
one must guess how to show them properly; for example some, like
FoundSet.find() seem to want to be plain text but with monospaced font
for ascii tables and preservation of newlines and spaces, some, like
JSDatabaseManager.saveData() seem to want to be pure HTML content, while
others (many) seem to want to preserve new lines and maybe spaces but
without the need to use monospaced font.

So this is an attempt to treat all those similarly both in SolEx.
tooltips and for markdown.