For SVY-19635: Foundsetpush issue
Some unit tests crashed; reasons:
- the foundsetID was different in a few cases because now more stuff is
sent in 1 go at startup
- serviceAPIs length was 2 instead of 1 for the same reason as above
- there was an exception being thrown because when the template JS
(tests do use ng1 form and template generator) did write the whole form
to JSON via WebFormUI.writeAllComponentsProperties, before actually
writing the JSON, in the super call, it would register the form to the
window and try to send the client side spec contents async, but because
this was happening outside a message-from-client-handling it would try
to send it right away to client and check for changes in the forms again
=> recursive call to WebFormUI.writeAllComponentsProperties and
exception. Made it so that sending client side spec contents for
components never tries to send the async message right away - as that
always happens when other messages will be sent afterwards (form will be
sent for sure after that) and will send the async call with them as
well.