For SVY-19635: Foundsetpush issue
Trying to make showing a form in tabpanel (or other similar components)
and showing a new form in main container both send data to client only
after form's onShow has been called without using invokeAndWait but
invokeLater, and with a custom event level instead. This way, even if
onShow does a sync call to component API, the form data can still be
normally sent after onShow, but sync call would not block the initial
form state from being sent to client (because the event level of that is
higher).

Purpose is to:
- send (initial) form data on form show to client after the onShow
handler executed, both for main form switches and to components that
show/hide forms
- allow sync calls to browser component API in onShow (code on client
needs to wait a bit for the next server event to execute on the event
thread if it doesn't yet find the form UI created)

Also added a new property type that allows one to set pageSize
(preferredViewportSize) of a foundset prop. directly from properties
view / solution code before show - and without the need to go to client
and call back setPreferredViewportSize... it will find it directly
server-side now before sending the initial data.