For SVY-19938: Regression on sample solution
If we get a message from client that is blocked on server because the
form is no longer visible, but it happens right after that form has been
hidden, do not report that as a WARN in the log - unless it is a
dataprovider change pushed from server.

For example if you have a text-field with onDataChange and onFocusLost
registered, you change the text from the field and then hit TAB,
onDataChange changes main form (so it hides previously shown form). Then
onFocusLost that was already sent from client happens - and that should
just be ignored without a warning being printed in the log.