SVY-19820 Handle bigints (java/javascript) to be able to send to the client (json doesnt directly support this)
added a utility method to parse a string to a double or BigDecimal
depending of the double can hold the value without loosing any
precision.

The DoublePropertyType toJSON will check for BigDecimals if the double
value can be extrated without loosing precission, if that is the case
then the value is written as a string (so javascript gets it as a
string)
fromJSON does it the other way around and will check if it can create a
double or it must be a BigDecimal when it gets a string.