Related to SVY-20327
A change that made parseConfig stricter in
FoundsetInitialPageSizePropertyType (it always expects the "for" to be
there as it makes no sense without it) made these tests fail with an
invizible exceptions.

Changed the test spec to make it valid + made the tests package reader
print out such exceptions instead of silently ignore them.