Fixes when the even thread is still executiong (long running stuff)
and then some kind of destroy happens. but then post() of that still
running stuff can recreated it but then suddenly its a different thread.

Introduced a EMPTY IEventExecutor.
This one is returned if createEventDispatcher(); is not returning
anything (the client is null or clientid == null , it shouldn't create
an event dispatcher anymore)

Als set it to the EMPTY value when the WebsocketSession is disposed() so
that a possible post after the fact doesn't suddenly create it again.